Added (unit)testing capability via pytest
