+++ main.py
@@ -1,5 +1,5 @@
 import telegram
-from creds import Telegram_API_Key
+from creds import creds
 
-bot = telegram.Bot(token=Telegram_API_Key())
+bot = telegram.Bot(token=creds.Telegram_API_Key())
 print(bot.get_me())
